Fetch status class has been made a singleton. Singleton meta class now has its own file. Changed string manipulation in fetch pool connection for something better. Added genesis path and url to fetch status class. Added comments.
